move implementation to protocols
