+++ api/src/opentrons/execute.py
@@ -19,7 +19,7 @@ from opentrons.config import IS_ROBOT, JUPYTER_NOTEBOOK_LABWARE_DIR
 from opentrons.protocol_api import (MAX_SUPPORTED_VERSION)
 from opentrons.protocols.execution import execute as execute_apiv2
 from opentrons import commands
-from opentrons.protocol_api.implementation.protocol_context import \
+from opentrons.protocols.implementation.protocol_context import \
     ProtocolContextImplementation
 from opentrons.protocols.parse import parse, version_from_string
 from opentrons.protocols.types import APIVersion, PythonProtocol
