+++ api/src/opentrons/protocols/implementation/interfaces/protocol_context.py
@@ -6,15 +6,15 @@ from typing import (Dict, List, Optional, Union)
 
 from opentrons import types
 from opentrons.hardware_control import API
-from opentrons.protocol_api.implementation.interfaces.versioned import \
+from opentrons.protocols.implementation.interfaces.versioned import \
     ApiVersioned
 from opentrons.protocol_api.labware import Labware
-from opentrons.protocol_api.module_geometry import ModuleGeometry
-from opentrons.protocol_api.geometry import Deck
+from opentrons.protocols.geometry.module_geometry import ModuleGeometry
+from opentrons.protocols.geometry.deck import Deck
 from opentrons.protocol_api.instrument_context import InstrumentContext
 from opentrons.protocol_api.module_contexts import ModuleContext, ModuleTypes
-from opentrons.protocol_api.util import AxisMaxSpeeds, \
-    HardwareManager
+from opentrons.protocols.api_support.util import (
+    AxisMaxSpeeds, HardwareManager)
 from opentrons_shared_data.labware.dev_types import LabwareDefinition
